Move non-efficiency stuff back into basic_metric, but still as static methods.
Later in this chain I'm going to want to call compute_request_state_metrics not in BasicMetric.
